feat: update for auto gen
