+++ rules/write_only/rule_test.go.tmpl
@@ -1,0 +1,66 @@
+package rules
+
+import (
+	"testing"
+
+	"github.com/terraform-linters/tflint-plugin-sdk/helper"
+)
+
+func Test_AwsWriteOnlyAttribute(t *testing.T) {
+	cases := []struct {
+		Name     string
+		Content  string
+		Expected helper.Issues
+		Fixed    string
+	}{
+        {{- range $name, $value := . }}
+		{{- range $kk, $writeOnly := $value }}
+		{
+			Name: "basic {{ $name }}",
+			Content: `
+resource "{{ $name }}" "test" {
+  {{ $writeOnly.OriginalAttribute }} = "test"
+}
+`,
+			Expected: helper.Issues{
+				{
+					Rule:    NewAwsWriteOnlyArgumentsRule(),
+					Message: `"{{ $writeOnly.OriginalAttribute }}" is a non-ephemeral attribute, which means this secret is stored in state. Please use write-only argument "{{ $writeOnly.WriteOnlyAlternative }}".`,
+				},
+			},
+			Fixed: `
+resource "{{ $name }}" "test" {
+  {{ $writeOnly.WriteOnlyAlternative }} = "test"
+}
+`,
+		},
+		{
+			Name: "everything is fine {{ $name }}",
+			Content: `
+resource "{{ $name }}" "test" {
+  {{ $writeOnly.WriteOnlyAlternative }} = "test"
+}
+`,
+			Expected: helper.Issues{},
+		},
+		{{- end -}}{{- end }}
+	}
+
+	rule := NewAwsWriteOnlyArgumentsRule()
+
+	for _, tc := range cases {
+		filename := "resource.tf"
+		runner := helper.TestRunner(t, map[string]string{filename: tc.Content})
+
+		if err := rule.Check(runner); err != nil {
+			t.Fatalf("Unexpected error occurred: %s", err)
+		}
+		helper.AssertIssuesWithoutRange(t, tc.Expected, runner.Issues)
+
+		want := map[string]string{}
+		if tc.Fixed != "" {
+			want[filename] = tc.Fixed
+		}
+		helper.AssertChanges(t, want, runner.Changes())
+	}
+}
