+++ rules/write_only/write_only.go
@@ -1,0 +1,3 @@
+//go:generate go run -tags generators ./generator/main.go
+
+package write_only
